03a/server_a/pyproject.toml
[tool.poetry]
name = "server_a"
version = "0.1.0"
description = "FastAPI server for data parsing (Server A)"
authors = ["HaDoCK <jako7096@stud.kea.dk>"]


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
pyyaml = "^6.0.2"
httpx = "^0.27.2"

[tool.poetry.scripts]
start = "uvicorn src.main:app --host 0.0.0.0 --port 8000"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.packages]]
include = "src"